+++ src/main/java/commoble/tubesreloaded/blocks/filter/FilterBlockEntity.java
@@ -17,8 +17,8 @@ import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.entity.BlockEntityType;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraftforge.common.capabilities.Capability;
+import net.minecraftforge.common.capabilities.ForgeCapabilities;
 import net.minecraftforge.common.util.LazyOptional;
-import net.minecraftforge.items.CapabilityItemHandler;
 import net.minecraftforge.items.IItemHandler;
 
 public class FilterBlockEntity extends BlockEntity
@@ -67,7 +67,7 @@ public class FilterBlockEntity extends BlockEntity
 	@Nonnull
 	public <T> LazyOptional<T> getCapability(@Nonnull Capability<T> cap, @Nullable Direction side)
 	{
-		if (cap == CapabilityItemHandler.ITEM_HANDLER_CAPABILITY)
+		if (cap == ForgeCapabilities.ITEM_HANDLER)
 		{
 			Direction output_dir = this.getBlockState().getValue(ShuntBlock.FACING);
 			if (side == output_dir.getOpposite())