+++ src/main/java/commoble/tubesreloaded/blocks/shunt/ShuntBlockEntity.java
@@ -10,8 +10,8 @@ import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.entity.BlockEntityType;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraftforge.common.capabilities.Capability;
+import net.minecraftforge.common.capabilities.ForgeCapabilities;
 import net.minecraftforge.common.util.LazyOptional;
-import net.minecraftforge.items.CapabilityItemHandler;
 
 public class ShuntBlockEntity extends BlockEntity
 {
@@ -45,7 +45,7 @@ public class ShuntBlockEntity extends BlockEntity
 	@Nonnull
 	public <T> LazyOptional<T> getCapability(@Nonnull Capability<T> cap, @Nullable Direction side)
 	{
-		if (cap == CapabilityItemHandler.ITEM_HANDLER_CAPABILITY)
+		if (cap == ForgeCapabilities.ITEM_HANDLER)
 		{
 			Direction output_dir = this.getBlockState().getValue(ShuntBlock.FACING);
 			return side == output_dir ? this.outputHandler.cast() : this.inputHandler.cast();